change the type of file_size and file_offset to off_t
git-svn-id: http://fastdfs.googlecode.com/svn@65 0d4c4929-bc51-0410-a161-b7c81ae5c018

+++ trunk/storage/storage_sync.h
@@ -31,7 +31,7 @@ typedef struct
 	int mark_fd;
 	int binlog_index;
 	int binlog_fd;
-	int binlog_offset;
+	off_t binlog_offset;
 	int scan_row_count;
 	int sync_row_count;
 	int last_write_row_count;
